api/requirements.txt
astroid==2.5.3
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==3.0.1
django-filter==2.0.0
django-heroku==0.0.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-filters==0.10.2
djangorestframework-xml==1.4.0
gunicorn==20.1.0
idna==2.10
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
